Organize the code into packages
