+++ docs/README.md
@@ -1,4 +1,4 @@
-# BugsBunny User Guide
+# bugsbunny.app.BugsBunny User Guide
 
 // Update the title above to match the actual product name
