+++ src/main/java/bugsbunny/commands/AddToDoCommand.java
@@ -1,3 +1,11 @@
+package bugsbunny.commands;
+
+import bugsbunny.app.Ui;
+import bugsbunny.storage.Storage;
+import bugsbunny.tasks.Task;
+import bugsbunny.tasks.TaskList;
+import bugsbunny.tasks.ToDo;
+
 import java.io.IOException;
 
 public class AddToDoCommand extends Command {
