+++ src/main/java/bugsbunny/commands/DeleteCommand.java
@@ -1,3 +1,11 @@
+package bugsbunny.commands;
+
+import bugsbunny.app.Ui;
+import bugsbunny.exception.BugsBunnyException;
+import bugsbunny.storage.Storage;
+import bugsbunny.tasks.Task;
+import bugsbunny.tasks.TaskList;
+
 import java.io.IOException;
 
 public class DeleteCommand extends Command {
