+++ src/main/java/bugsbunny/commands/UnmarkCommand.java
@@ -1,3 +1,10 @@
+package bugsbunny.commands;
+
+import bugsbunny.app.Ui;
+import bugsbunny.exception.BugsBunnyException;
+import bugsbunny.storage.Storage;
+import bugsbunny.tasks.TaskList;
+
 import java.io.IOException;
 
 public class UnmarkCommand extends Command {
