+++ src/main/java/bugsbunny/exception/BugsBunnyException.java
@@ -1,3 +1,5 @@
+package bugsbunny.exception;
+
 public class BugsBunnyException extends Exception {
     public BugsBunnyException(String message) {
         super(message);
