+++ src/main/java/bugsbunny/tasks/Event.java
@@ -1,3 +1,7 @@
+package bugsbunny.tasks;
+
+import bugsbunny.parsers.DateTimeParser;
+
 import java.time.LocalDateTime;
 
 public class Event extends Task {
